Added support for sending the contexts field
Previously, the contexts field could be sent but nothing was done with it, because it was being sent as a string and not as an array.

This update JSON-decodes the contexts field from the command line so that when it is extracted by the IET, it is properly utilized in the incident.