Support Pagerduty V2 PD-CEF fields
Per PR 119 on pdagent, this extends the Pagerduty V2 API support to
pdagent-integrations by:

- Sending through the required PD-CEF fields (these need to be sent to
pdagent even if using V1 api per the current status of PR 119, even if
not actually used for V1 api call)
- Adding a command line flag to select the API version (defaults to V1)
- Adds default mappings for certain PD-CEF fields from Sensu check results
- Allows an optional argument to customise the mappings for PD-CEF
fields from Sensu check results. That works as follows:

        --event-map "event_source,check.name"

    Comma separated and then the field we want from the check result in
    dot notation. It's fairly simplistic, not going to support fields
    that actually use dots in their name (but probably pretty rare in
    Sensu).

    That argument can be supplied multiple times for the different
    fields.

It's _fairly_ well error handled (not perfectly) so that if a mapping fails
(e.g. because that field doesn't exist in the Sensu check result)
then it will just set that PD-CEF to an empty string

Have tried to copy command structure of PR 119 as much as is applicable.

Have not looked at tests yet.

References: https://github.com/PagerDuty/pdagent/pull/119