new optional argument as record separator for multi-line text support in zabbix-integration
An optional fourth argument is introduced as the record separator prefix 'eor_sep_prefix' for the parser of alert text 'body_str'.
The new record separator is defined as 'eor_sep_prefix' + "<newline>".
We can use any comment text of multi-line for a pagerduty alert even if the alert text contains ":" because the record separator can be set to a rare word, such as "@#@%@<newline>".
The default value of optional fourth argument 'eor_sep_prefix' is an empty string to preserve the backward compatibility of script.